BackEnd/src/main/java/com/Airline/Security/Jwt/JwtTokenUtil.java
package com.Airline.Security.Jwt;

import io.jsonwebtoken.*;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.security.core.Authentication;
import org.springframework.stereotype.Component;
import org.springframework.security.core.GrantedAuthority;


import javax.crypto.spec.SecretKeySpec;
import java.security.Key;
import java.util.Collections;
import java.util.Date;
import java.util.List;
import java.util.stream.Collectors;

@Slf4j
@Component
public class JwtTokenUtil {

    private final Key key;

    @Value("${jwt.expirationMs}")
    int jwtExpirationMs;

    public JwtTokenUtil(@Value("fvtpE9CIsCfnHJx7tpIXeB3chpKrIImW2XpfDB0q+J4=") String secret) {
        this.key = new SecretKeySpec(secret.getBytes(), SignatureAlgorithm.HS512.getJcaName());
    }

    public String generateToken(Authentication authentication) {
        Date now = new Date();
        Date expiryDate = new Date(now.getTime() + jwtExpirationMs);

        Claims claims = Jwts.claims().setSubject(authentication.getName());
        claims.put("roles", authentication.getAuthorities().stream()
                .map(GrantedAuthority::getAuthority)
                .collect(Collectors.toList()));

        // Log the roles being set into the JWT for debugging
        System.out.println("Roles added to JWT: " + claims.get("roles"));

        return Jwts.builder()
                .setClaims(claims)
                .setIssuedAt(now)
                .setExpiration(expiryDate)
                .signWith(key)
                .compact();
    }



    public List<String> getRolesFromToken(String token) {
        try {
            Claims claims = Jwts.parser()
                    .setSigningKey(key)
                    .parseClaimsJws(token)
                    .getBody();
            return claims.get("roles", List.class); // Assuming roles are stored as a List of String
        } catch (JwtException | IllegalArgumentException e) {
            log.error("Failed to extract roles from JWT", e);
            return Collections.emptyList(); // Return an empty list on failure
        }
    }
    public String getUsernameFromToken(String token) {
        return Jwts.parser()
                .setSigningKey(key)
                .parseClaimsJws(token)
                .getBody()
                .getSubject();
    }

    public boolean validateToken(String authToken) {
        try {
            Jwts.parser()
                    .setSigningKey(key)
                    .parseClaimsJws(authToken);
            return true;
        } catch (JwtException | IllegalArgumentException e) {
            // Log the exception details, e.g., log.error("Token validation error", e);
        }
        return false;
    }
}
